cleaned up code (nsfw)

+++ Classes/Items/DropMinorItem.cs
@@ -1,7 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace CSE3902_Game_Sprint0.Classes.Items
 {
@@ -18,7 +16,7 @@ namespace CSE3902_Game_Sprint0.Classes.Items
         public void Execute()
         {
             var random = new Random();
-            int randomNumber = (int)random.Next(100);
+            int randomNumber = random.Next(100);
 
             if (randomNumber <= 5)
             {
